requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
mssql-django==1.0
pyodbc==4.0.32
pytz==2021.1
sqlparse==0.4.2
